issue #7 : Shade guava
Guava regularly bumps its major version and includes method removals in the process of doing so, which fits with Semantic Versioning, but which limits reuse of libraries that externally depend on it. This removes that obstacle by shading/removing Guava as an external dependency, so it can be upgraded locally without affecting others.

Also updates Jackson, which is already shaded, Apache Commons libraries (do not need shading as they have a much longer maintenance cycle), and Slf4J (Does not need shading as it has a very reputable history in terms of backwards compatibility)

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>