Use GNU sort in more cases, but only if it's new enough.
This helps when multiple files need to be sorted, since there's more chance
the same method will be used.

"New" (i.e. not the really old one on Mac OS) GNU sort supports -k1,1 format
for a sort column specification, which is necessary when the sorted file
contains values like

10,xx
100,xx

to ensure this order.

Resolves POR-2464.